Finish creating a vertical navbar
Add an user account image (for now to public images directory). Create
an account section inside the navbar and make dropdown for it. Also
make dropdown with subpages for Flashcard and Test (navigation items).

Add custom styles for content, footer and vertical navbar and modify
some styles but it's not refactoring yet. Make a code a little bit more
accessible writing WAI-ARIA stuff.