Refactor sidebar and make flexible all elements
Modify and reduce count of custom styles for the whole side. Add some
new styles (in the same files) especially for scrollbar and sidebar.

Modify and make more accessible an HTML structure for twig base
template and twig homepage. Also change and optimise Bootstrap
classes for the elements on the page.

Turn on jQuery support in the webpack config file and make jQuery
global inside assets.