Add "show a word translation" function
Add FlashcardController::showFlashcardTranslation method inside the
FlashcardController that takes a id parameter which is a number of
the choosen card. Write the route for this function.

Inside the main twig template (main.html.twig) modify the code by
adding condition which check if id of a card exist in URI (id exists
when user choose some card or when write the URI with that id by
oneself). If YES - display only translation of specific flashcard and
display the rest of the cards by default (foreign word + pronunciation
and example sentence if choosen in top bar), If NO - display everything
by default.

Move CSS styles for flashcards to another file (_flashcard.scss) and
directory (/css/components) and modify the min height for card.